+++ load-package.php
@@ -2,7 +2,7 @@
 /**
  * This file is designed to be used to load as package NOT a WP plugin!
  *
- * @version 3.0.3
+ * @version 3.1.0-beta.1
  * @package CoCart JWT Authentication
  */